% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqnumber.R
\name{print.pqnumber}
\alias{print.pqnumber}
\title{Print a pqnumber in a good manner}
\usage{
\method{print}{pqnumber}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{a pqnumber}
}
\description{
Print a pqnumber in a good manner
}
